# [O11Y-197] optimize orderOf algorithms on x86 with AVX-512 support