--- a/internal/bits/order_test.go
+++ b/internal/bits/order_test.go
@@ -202,7 +202,7 @@ func TestOrderOfInt64(t *testing.T) {
 	}
 
 	if !check(values) {
-		t.Error("failed due to not checking the connection between sequences of of 8 elements")
+		t.Error("failed due to not checking the connection between sequences of 8 elements")
 	}
 }