--- a/internal/bits/order_test.go
+++ b/internal/bits/order_test.go
@@ -236,7 +236,7 @@ func TestOrderOfUint32(t *testing.T) {
 	}
 
 	if !check(values) {
-		t.Error("failed due to not checking the connection between sequences of of 16 elements")
+		t.Error("failed due to not checking the connection between sequences of 16 elements")
 	}
 }