--- a/internal/bits/order_test.go
+++ b/internal/bits/order_test.go
@@ -270,7 +270,7 @@ func TestOrderOfUint64(t *testing.T) {
 	}
 
 	if !check(values) {
-		t.Error("failed due to not checking the connection between sequences of of 8 elements")
+		t.Error("failed due to not checking the connection between sequences of 8 elements")
 	}
 }