Continuing progress on the work initiated in #59 and #59, this PR adds optimizations for other routines operating on arrays of fixed-size values:

```
name                    old time/op    new time/op     delta
OrderOfBool/4KiB          90.7ns ± 0%     90.5ns ± 0%    -0.21%  (p=0.004 n=10+10)
OrderOfBool/256KiB        6.65µs ± 0%     6.65µs ± 0%      ~     (p=0.646 n=10+9)
OrderOfBool/2048KiB       66.7µs ± 5%     67.1µs ± 4%      ~     (p=0.631 n=10+10)
OrderOfInt32/4KiB         1.40µs ± 0%     0.18µs ± 0%   -87.16%  (p=0.000 n=9+9)
OrderOfInt32/256KiB       92.9µs ± 1%     21.9µs ± 1%   -76.45%  (p=0.000 n=10+10)
OrderOfInt32/2048KiB       738µs ± 2%      275µs ± 5%   -62.72%  (p=0.000 n=10+10)
OrderOfInt64/4KiB         1.40µs ± 1%     0.37µs ± 0%   -73.52%  (p=0.000 n=10+8)
OrderOfInt64/256KiB        104µs ± 3%       65µs ± 2%   -37.19%  (p=0.000 n=10+9)
OrderOfInt64/2048KiB      1.31ms ±15%     0.55ms ± 5%   -57.59%  (p=0.000 n=10+10)
OrderOfUint32/4KiB        1.40µs ± 0%     0.18µs ± 1%   -87.16%  (p=0.000 n=10+10)
OrderOfUint32/256KiB      94.5µs ± 1%     23.6µs ± 1%   -75.02%  (p=0.000 n=10+10)
OrderOfUint32/2048KiB      730µs ± 1%      268µs ± 8%   -63.24%  (p=0.000 n=9+10)
OrderOfUint64/4KiB        1.40µs ± 1%     0.37µs ± 1%   -73.46%  (p=0.000 n=9+10)
OrderOfUint64/256KiB       105µs ± 2%       65µs ± 8%   -37.68%  (p=0.000 n=10+10)
OrderOfUint64/2048KiB     1.31ms ±21%     0.55ms ± 7%   -57.55%  (p=0.000 n=10+10)
OrderOfFloat32/4KiB       1.90µs ± 0%     0.18µs ± 1%   -90.59%  (p=0.000 n=9+9)
OrderOfFloat32/256KiB      122µs ± 0%       24µs ± 2%   -80.43%  (p=0.000 n=9+10)
OrderOfFloat32/2048KiB    1.01ms ± 2%     0.27ms ± 5%   -73.24%  (p=0.000 n=10+10)
OrderOfFloat64/4KiB       1.90µs ± 0%     0.37µs ± 1%   -80.46%  (p=0.000 n=6+10)
OrderOfFloat64/256KiB      128µs ± 1%       65µs ± 5%   -49.27%  (p=0.000 n=10+10)
OrderOfFloat64/2048KiB    1.46ms ±21%     0.56ms ± 5%   -61.46%  (p=0.000 n=10+10)
OrderOfBytes/4KiB          851ns ± 1%      772ns ± 0%    -9.20%  (p=0.000 n=10+10)
OrderOfBytes/256KiB       53.7µs ± 0%     49.2µs ± 1%    -8.39%  (p=0.000 n=10+10)
OrderOfBytes/2048KiB       434µs ± 0%      397µs ± 1%    -8.46%  (p=0.000 n=9+10)

name                    old speed      new speed       delta
OrderOfBool/4KiB        45.2GB/s ± 0%   45.3GB/s ± 0%    +0.21%  (p=0.005 n=10+10)
OrderOfBool/256KiB      39.4GB/s ± 0%   39.4GB/s ± 0%      ~     (p=0.661 n=10+9)
OrderOfBool/2048KiB     31.5GB/s ± 5%   31.3GB/s ± 5%      ~     (p=0.631 n=10+10)
OrderOfInt32/4KiB       2.93GB/s ± 0%  22.86GB/s ± 0%  +678.99%  (p=0.000 n=9+9)
OrderOfInt32/256KiB     2.82GB/s ± 1%  11.98GB/s ± 1%  +324.56%  (p=0.000 n=10+10)
OrderOfInt32/2048KiB    2.84GB/s ± 2%   7.63GB/s ± 5%  +168.39%  (p=0.000 n=10+10)
OrderOfInt64/4KiB       2.93GB/s ± 1%  11.07GB/s ± 0%  +277.71%  (p=0.000 n=10+8)
OrderOfInt64/256KiB     2.53GB/s ± 3%   4.02GB/s ± 2%   +59.21%  (p=0.000 n=10+9)
OrderOfInt64/2048KiB    1.62GB/s ±17%   3.79GB/s ± 5%  +134.27%  (p=0.000 n=10+10)
OrderOfUint32/4KiB      2.93GB/s ± 0%  22.85GB/s ± 1%  +678.95%  (p=0.000 n=10+10)
OrderOfUint32/256KiB    2.77GB/s ± 1%  11.10GB/s ± 1%  +300.36%  (p=0.000 n=10+10)
OrderOfUint32/2048KiB   2.87GB/s ± 1%   7.83GB/s ± 8%  +172.67%  (p=0.000 n=9+10)
OrderOfUint64/4KiB      2.93GB/s ± 1%  11.03GB/s ± 1%  +276.79%  (p=0.000 n=9+10)
OrderOfUint64/256KiB    2.50GB/s ± 2%   4.02GB/s ± 7%   +60.71%  (p=0.000 n=10+10)
OrderOfUint64/2048KiB   1.63GB/s ±21%   3.79GB/s ± 7%  +131.79%  (p=0.000 n=10+10)
OrderOfFloat32/4KiB     2.15GB/s ± 0%  22.87GB/s ± 1%  +962.98%  (p=0.000 n=9+9)
OrderOfFloat32/256KiB   2.16GB/s ± 0%  11.02GB/s ± 2%  +410.94%  (p=0.000 n=9+10)
OrderOfFloat32/2048KiB  2.07GB/s ± 2%   7.74GB/s ± 5%  +273.99%  (p=0.000 n=10+10)
OrderOfFloat64/4KiB     2.15GB/s ± 0%  11.01GB/s ± 1%  +411.88%  (p=0.000 n=10+10)
OrderOfFloat64/256KiB   2.04GB/s ± 1%   4.03GB/s ± 5%   +97.21%  (p=0.000 n=10+10)
OrderOfFloat64/2048KiB  1.45GB/s ±24%   3.72GB/s ± 5%  +156.28%  (p=0.000 n=10+10)
OrderOfBytes/4KiB       4.82GB/s ± 1%   5.30GB/s ± 0%   +10.12%  (p=0.000 n=10+10)
OrderOfBytes/256KiB     4.89GB/s ± 0%   5.33GB/s ± 1%    +9.16%  (p=0.000 n=10+10)
OrderOfBytes/2048KiB    4.84GB/s ± 0%   5.28GB/s ± 1%    +9.24%  (p=0.000 n=9+10)
```